requirements.txt
alembic==1.0.0
aniso8601==3.0.2
astroid==2.0.4
atomicwrites==1.2.1
attrs==18.2.0
autoenv==1.0.0
backcall==0.1.0
bcrypt==3.1.4
blinker==1.4
cached-property==1.5.1
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
click==6.7
codecov==2.0.15
colorama==0.3.9
coverage==4.5.1
coveralls==1.5.0
decorator==4.3.0
docker==3.5.0
docker-compose==1.22.0
docker-pycreds==0.3.0
dockerpty==0.4.1
docopt==0.6.2
flake8==3.5.0
Flask==1.0.2
Flask-Admin==1.5.2
Flask-API==1.0
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.6
Flask-HTTPAuth==3.2.4
Flask-JWT==0.3.2
Flask-JWT-Extended==3.13.0
Flask-Login==0.4.1
Flask-Mail==0.9.1
Flask-MicroServices==0.34.5
Flask-Migrate==2.2.1
Flask-MongoAlchemy==0.7.2
Flask-PyMongo==2.1.0
Flask-RESTful==0.3.6
Flask-Session==0.3.1
Flask-SocketIO==3.0.1
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
freezegun==0.3.10
get==1.0.3
gunicorn==19.9.0
idna==2.6
ipdb==0.11
ipython==6.5.0
ipython-genutils==0.2.0
isort==4.3.4
itsdangerous==0.24
jedi==0.12.1
Jinja2==2.10
jsonschema==2.6.0
lazy-object-proxy==1.3.1
Mako==1.0.7
Markdown==2.6.11
MarkupSafe==1.0
mccabe==0.6.1
MongoAlchemy==0.19
mongoengine==0.15.3
more-itertools==4.3.0
mpesa-api==0.1.8
nose==1.3.7
numpy==1.15.0
parso==0.3.1
passlib==1.7.1
pickleshare==0.7.4
pipenv==2018.7.1
pluggy==0.7.1
post==1.0.2
progressbar2==3.38.0
prompt-toolkit==1.0.15
psycopg2==2.7.5
public==1.0.3
pudb==2018.1
py==1.6.0
pycodestyle==2.3.1
pycparser==2.18
pyflakes==1.6.0
Pygments==2.2.0
PyJWT==1.4.2
pylint==2.1.1
pymongo==2.8.1
pypiwin32==223
pytest==3.7.4
pytest-cov==2.6.0
pytest-freezegun==0.2.0
pytest-html==1.19.0
pytest-metadata==1.7.0
python-dateutil==2.7.3
python-dotenv==0.9.1
python-editor==1.0.3
python-engineio==2.2.0
python-socketio==2.0.0
python-utils==2.3.0
pytz==2018.5
pywin32==223
PyYAML==3.13
query-string==1.0.2
redis==2.10.6
request==1.0.2
requests==2.18.4
simplegeneric==0.8.1
six==1.11.0
SQLAlchemy==1.2.12
texttable==0.9.1
traitlets==4.3.2
urllib3==1.22
urwid==2.0.1
versiontools==1.9.1
virtualenv==16.0.0
virtualenv-clone==0.3.0
virtualenvwrapper-win==1.2.5
waitress==1.1.0
wcwidth==0.1.7
websocket-client==0.53.0
Werkzeug==0.14.1
wrapt==1.10.11
WTForms==2.2.1
